LineChatBotApi/requirements.txt
Flask==1.1.2
line-bot-sdk==1.17.0
gunicorn==20.0.4
urllib3==1.25.9
Werkzeug==1.0.1
opentracing>=2.0,<2.4
pymongo>=3.1
six
dnspython